+++ DMHelper/src/battledialogmodelmonsterclass.h
@@ -24,7 +24,7 @@ public:
 
     // Local
     virtual BattleDialogModelMonsterClass* clone() const override;
-    virtual int getSizeFactor() const override;
+    virtual qreal getSizeFactor() const override;
     virtual int getSizeCategory() const override;
 
     virtual int getStrength() const override;
